third_party/blink/web_tests/flag-specific/highdpi/fast/events/offsetX-offsetY-svg-expected.txt

This is a testharness.js-based test.
FAIL MouseEvent.offsetX/offsetY for an SVG element is always relative to the outermost <svg> CSS layout box assert_equals: offsetX expected 80 but got 71
Harness: the test ran to completion.
